fix(evohome): fan mode switch commands were ignored on signed-char platforms
WriteToHardware dispatches on pdata[1], a plain char. The fan mode
selector added with the RAMSES-II HVAC support is a pTypeGeneralSwitch
(0xF4) packet, and on platforms where char is signed that byte reads as
-12, so its case label never matched and the command was silently
dropped. Only ARM builds (unsigned char) worked. Cast the type byte to
uint8_t so the dispatch is correct everywhere; the EvoHome packet types
0x44-0x47 were unaffected.

+++ hardware/EvohomeRadio.cpp
@@ -224,7 +224,7 @@ bool CEvohomeRadio::WriteToHardware(const char* pdata, const unsigned char lengt
 {
 	if (!pdata)
 		return false;
-	switch (pdata[1])
+	switch (static_cast<uint8_t>(pdata[1]))
 	{
 	case pTypeEvohome:
 	{
